requirements.txt
openai-whisper
torch
ffmpeg-python
python-dotenv
pydub
numpy
moviepy